+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/DeclarationOfExpenditureCertificate/DeclarationOfExpenditureCertificate.cshtml.cs
@@ -8,7 +8,6 @@ using Dfe.Complete.Services;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.ComponentModel.DataAnnotations;
 using static Dfe.Complete.Services.DateRangeValidationService;
 
 namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.DeclarationOfExpenditureCertificate