+++ CHANGELOG.md
@@ -16,8 +16,8 @@ Security in case of vulnerabilities.
 
 ## Unreleased
 
-### Added
 
+### Added
 - Added `Redact and send document` task for both conversion and transfer projects.
 - Added `Receivee declaration of expenditure certificate` task for conversion and transfer projects.
 
@@ -30,6 +30,7 @@ Security in case of vulnerabilities.
 
 ### Fixed
 - Complete project post no longer throws antiforgery error
+- Correct wording on handover complete confirmation page
 
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-03.710...main) for everything awaiting release