+++ CHANGELOG.md
@@ -16,9 +16,10 @@ Security in case of vulnerabilities.
 
 ## Unreleased
 
-### Added
 
+### Added
 - Added `Redact and send document` task for both conversion and transfer projects.
+- Added `Confirm proposed capacity of the academy` task for conversion project.
 
 ### Security
 - prevent inactive users from signing in
@@ -29,6 +30,7 @@ Security in case of vulnerabilities.
 
 ### Fixed
 - Complete project post no longer throws antiforgery error
+- Correct wording on handover complete confirmation page
 
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-03.710...main) for everything awaiting release